pyproject.toml
[project]
name = "argo"
version = "0.1.0"
description = "Argo — The Argonauts SOC Platform (Orpheus CTI Agent)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.5",
  "rich",
  "pydantic>=2",
  "psycopg[binary]",
  "faiss-cpu",
  "pymupdf",
  "python-dotenv",
  "rank-bm25",
  "openai>=1.40.0",
  "tqdm",
  "tenacity",
  "langgraph>=0.2.21",        # NEW
  "langchain-core>=0.2.39",   # lightweight core runtime LangGraph expects
  "structlog>=24.1.0",        # structured logging for audit trails
]

[project.optional-dependencies]
ocr = ["pytesseract", "Pillow"]
dev = ["ruff", "pytest", "mypy"]

[project.scripts]
argo = "argo.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["argo"]  # install the argo package (with subpackages)

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
warn_unused_ignores = true
